+++ tests/model_serving/model_server/authentication/test_kserve_token_authentication.py
@@ -10,7 +10,7 @@ pytestmark = pytest.mark.usefixtures("skip_if_no_authorino_operator", "valid_aws
 
 
 @pytest.mark.parametrize(
-    "model_namespace, s3_models_storage_uri",
+    "model_namespace, s3_models_storage_uri, ",
     [
         pytest.param(
             {"name": "kserve-token-authentication"},
